pyproject.toml
[tool.poetry]
name = "bds-agent"
version = "0.1.0"
description = "CLI for BDS agents: signup, credits, run, MCP (stdio), LLM backends"
authors = ["Powerloom"]
readme = "README.md"
packages = [{ include = "bds_agent", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.0"
rich = ">=13.0,<15"
typer = "^0.15.4"
pympp = { version = "^0.5.4", extras = ["tempo"] }
pyyaml = "^6.0.3"
pydantic = "^2.12.5"
mcp = ">=1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
ruff = "^0.9.0"

[tool.poetry.scripts]
bds-agent = "bds_agent.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
